Moved the priceCheck namespace to Prices
