+++ src/Sidekick/App.xaml
@@ -2,7 +2,7 @@
     x:Class="Sidekick.App"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:converters="clr-namespace:Sidekick.Windows.PriceCheck.Converters"
+    xmlns:converters="clr-namespace:Sidekick.Windows.Prices.Converters"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     ShutdownMode="OnExplicitShutdown"
