+++ src/Sidekick/Windows/Prices/OverlayController.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Diagnostics;
 using System.Linq;
 using System.Threading.Tasks;
 using System.Windows;
@@ -12,43 +11,32 @@ using Sidekick.Business.Trades;
 using Sidekick.Business.Trades.Results;
 using Sidekick.Core.Loggers;
 using Sidekick.Core.Natives;
-using Sidekick.Core.Settings;
-using Application = System.Windows.Application;
 
-namespace Sidekick.Windows.PriceCheck
+namespace Sidekick.Windows.Prices
 {
     public class OverlayController : IDisposable
     {
         private readonly ITradeClient tradeClient;
-        private readonly INativeProcess nativeProcess;
         private readonly INativeClipboard clipboard;
-        private readonly INativeCursor nativeCursor;
         private readonly ILogger logger;
         private readonly IItemParser itemParser;
         private readonly IPoeNinjaCache poeNinjaCache;
-        private readonly SidekickSettings settings;
         private readonly OverlayWindow overlayWindow;
 
         public OverlayController(
             ITradeClient tradeClient,
-            INativeProcess nativeProcess,
             IKeybindEvents events,
             INativeClipboard clipboard,
-            INativeCursor nativeCursor,
             ILogger logger,
             IItemParser itemParser,
             IPoeNinjaCache poeNinjaCache,
-            SidekickSettings settings,
             IServiceProvider serviceProvider)
         {
             this.tradeClient = tradeClient;
-            this.nativeProcess = nativeProcess;
             this.clipboard = clipboard;
-            this.nativeCursor = nativeCursor;
             this.logger = logger;
             this.itemParser = itemParser;
             this.poeNinjaCache = poeNinjaCache;
-            this.settings = settings;
 
             overlayWindow = serviceProvider.GetService<OverlayWindow>();
             overlayWindow.MouseDown += Window_OnHandleMouseDrag;
@@ -56,7 +44,6 @@ namespace Sidekick.Windows.PriceCheck
 
             events.OnCloseWindow += OnCloseWindow;
             events.OnPriceCheck += OnPriceCheck;
-            events.OnMouseClick += MouseClicked;
         }
 
         public bool IsDisplayed => overlayWindow.IsDisplayed;
@@ -85,40 +72,7 @@ namespace Sidekick.Windows.PriceCheck
         /// </summary>
         public void Open()
         {
-            var scale = 96f / nativeProcess.ActiveWindowDpi;
-            var cursorPosition = nativeCursor.GetCursorPosition();
-            var xScaled = (int)(cursorPosition.X * scale);
-            var yScaled = (int)(cursorPosition.Y * scale);
-
-            EnsureBounds(xScaled, yScaled, scale);
             Show();
-        }
-
-        public Point GetOverlayPosition()
-        {
-            Debug.Assert(Application.Current.Dispatcher != null, "Application.Current.Dispatcher != null");
-            return Application.Current.Dispatcher.Invoke(() => new Point(overlayWindow.Left, overlayWindow.Top));
-        }
-
-        public Size GetOverlaySize()
-        {
-            return new Size(overlayWindow.ActualWidth, overlayWindow.ActualHeight);
-        }
-
-        /// <summary>
-        /// Ensures that the window stays within width and height of the display.
-        /// </summary>
-        private void EnsureBounds(int desiredX, int desiredY, float scale)
-        {
-            var screenRect = nativeProcess.GetScreenDimensions();
-
-            var xMidScaled = (screenRect.X + (screenRect.Width / 2)) * scale;
-            var yMidScaled = (screenRect.Y + (screenRect.Height / 2)) * scale;
-
-            var positionX = desiredX + (desiredX < xMidScaled ? overlayWindow.Padding.Left : -overlayWindow.Width - overlayWindow.Padding.Left);
-            var positionY = desiredY + (desiredY < yMidScaled ? overlayWindow.Padding.Top : -overlayWindow.Height - overlayWindow.Padding.Top);
-
-            overlayWindow.SetWindowPosition((int)positionX, (int)positionY);
         }
 
         /// <summary>
@@ -178,22 +132,6 @@ namespace Sidekick.Windows.PriceCheck
             }
 
             return false;
-        }
-
-        private Task MouseClicked(int x, int y)
-        {
-            if (!IsDisplayed || !settings.CloseOverlayWithMouse) return Task.CompletedTask;
-
-            var overlayPos = GetOverlayPosition();
-            var overlaySize = GetOverlaySize();
-
-            if (x < overlayPos.X || x > overlayPos.X + overlaySize.Width
-                                 || y < overlayPos.Y || y > overlayPos.Y + overlaySize.Height)
-            {
-                Hide();
-            }
-
-            return Task.CompletedTask;
         }
     }
 }
