+++ src/Sidekick/Windows/Prices/OverlayView.xaml
@@ -1,25 +1,22 @@
 <Window
-    x:Class="Sidekick.Windows.PriceCheck.OverlayWindow"
+    x:Class="Sidekick.Windows.Prices.OverlayWindow"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:VM="clr-namespace:Sidekick.Windows.PriceCheck.ViewModels"
     xmlns:loc="clr-namespace:Sidekick.Windows"
     xmlns:resx="clr-namespace:Sidekick.Localization.PriceCheck;assembly=Sidekick.Localization"
-    loc:Translation.ResourceManager="{x:Static resx:PriceCheckResources.ResourceManager}"
-    SizeToContent="WidthAndHeight">
+    loc:Translation.ResourceManager="{x:Static resx:PriceCheckResources.ResourceManager}">
     <Window.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="/Styles/Main.xaml" />
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
+        <ResourceDictionary Source="/Styles/Main.xaml" />
     </Window.Resources>
 
     <Window.Style>
         <Style BasedOn="{StaticResource WindowOverlay}" TargetType="Window" />
     </Window.Style>
 
-    <DockPanel Margin="{StaticResource Spacer}">
+    <DockPanel
+        Width="600"
+        Height="320"
+        Margin="{StaticResource Spacer}">
         <Label
             Margin="{StaticResource SpacerBottom}"
             VerticalAlignment="Center"
@@ -27,7 +24,7 @@
             DockPanel.Dock="Top"
             FontSize="16"
             FontWeight="Bold"
-            Foreground="{Binding queryResult.Item.Rarity, Converter={StaticResource RarityToColorConverter}}" />
+            Foreground="{Binding queryResult.Item, Converter={StaticResource RarityToColorConverter}}" />
 
         <Grid DockPanel.Dock="Bottom">
             <Grid.ColumnDefinitions>
@@ -91,7 +88,7 @@
                                     <DockPanel>
                                         <TextBlock Text="{Binding Price.Amount}" />
                                         <TextBlock Text=" " />
-                                        <Image Height="20">
+                                        <Image Height="20" ToolTip="{Binding Price.Currency}">
                                             <Image.Source>
                                                 <BitmapImage DecodePixelWidth="20" UriSource="{Binding Price.CurrencyUrl}" />
                                             </Image.Source>
